+++ game_patch/hud/hud.cpp
@@ -447,4 +447,5 @@ void hud_apply_patches()
     message_log_apply_patch();
     hud_world_apply_patch();
     hud_colors_apply_patch();
+    hud_scale_apply_patch();
 }